+++ packages/antd/src/index.ts
@@ -1,6 +1,16 @@
 export * from "./components/application"
+export { type ApplicationHeaderProps } from "./components/application/fragments/header"
+export {
+	type ApplicationConfig,
+	type ApplicationConfigBase,
+	type ApplicationConfigProviderProps,
+	type ApplicationContextValue,
+	type ApplicationSidebar,
+	type ApplicationState,
+} from "./components/application/context"
 export * from "./components/icon-select"
 export * from "./components/layout"
+export { type LayoutSidebarProps } from "./components/layout/sidebar"
 export * from "./components/page"
 export * from "./components/with-overlay"
 export * from "./i18n"
